+++ src/StdMeshersGUI/StdMeshersGUI_ObjectReferenceParamWdg.h
@@ -68,16 +68,16 @@ public:
   void SetObjects(SMESH::string_array_var& objEntries);
 
   template<class TInterface> 
-    typename TInterface::_var_type GetObject(unsigned i=0) const {
+    typename TInterface::_var_type GetObject(size_t i=0) const {
     if ( IsObjectSelected(i) ) return TInterface::_narrow(myObjects[i]);
     return TInterface::_nil();
   }
 
-  int NbObjects() const { return myObjects.size(); }
+  size_t NbObjects() const { return myObjects.size(); }
 
   QString GetValue() const { return myParamValue; }
 
-  bool IsObjectSelected(unsigned i=0) const
+  bool IsObjectSelected(size_t i=0) const
   { return i < myObjects.size() && !CORBA::is_nil(myObjects[i]); }
 
   /*!
